+++ pylops/signalprocessing/FFTND.py
@@ -1,128 +1,6 @@
-import logging
-import warnings
-
 import numpy as np
-from numpy.core.multiarray import normalize_axis_index
 
-from pylops import LinearOperator
-from pylops.utils.backend import get_complex_dtype, get_real_dtype
-
-from ._BaseFFTs import _raise_on_wrong_dtype, _value_or_list_like_to_array
-
-logging.basicConfig(format="%(levelname)s: %(message)s", level=logging.WARNING)
-
-
-class _BaseFFTND(LinearOperator):
-    """Base class for N-dimensional fast Fourier Transform"""
-
-    def __init__(
-        self,
-        dims,
-        dirs=None,
-        nffts=None,
-        sampling=1.0,
-        real=False,
-        ifftshift_before=False,
-        fftshift_after=False,
-        dtype="complex128",
-    ):
-        self.dims = _value_or_list_like_to_array(dims)
-        _raise_on_wrong_dtype(self.dims, np.integer, "dims")
-
-        self.ndim = len(self.dims)
-
-        dirs = _value_or_list_like_to_array(dirs)
-        _raise_on_wrong_dtype(dirs, np.integer, "dirs")
-        self.dirs = np.array([normalize_axis_index(d, self.ndim) for d in dirs])
-        self.ndirs = len(self.dirs)
-        if self.ndirs != len(np.unique(self.dirs)):
-            warnings.warn(
-                "At least one direction is repeated. This may cause unexpected results."
-            )
-
-        nffts = _value_or_list_like_to_array(nffts, repeat=self.ndirs)
-        if len(nffts[np.equal(nffts, None)]) > 0:  # Found None(s) in nffts
-            nffts[np.equal(nffts, None)] = np.array(
-                [self.dims[d] for d, n in zip(dirs, nffts) if n is None]
-            )
-            nffts = nffts.astype(self.dims.dtype)
-        self.nffts = nffts
-        _raise_on_wrong_dtype(self.nffts, np.integer, "nffts")
-
-        sampling = _value_or_list_like_to_array(sampling, repeat=self.ndirs)
-        if np.issubdtype(sampling.dtype, np.integer):  # Promote to float64 if integer
-            sampling = sampling.astype(np.float64)
-        self.sampling = sampling
-        _raise_on_wrong_dtype(self.sampling, np.floating, "sampling")
-
-        self.ifftshift_before = _value_or_list_like_to_array(
-            ifftshift_before, repeat=self.ndirs
-        )
-        _raise_on_wrong_dtype(self.ifftshift_before, bool, "ifftshift_before")
-
-        self.fftshift_after = _value_or_list_like_to_array(
-            fftshift_after, repeat=self.ndirs
-        )
-        _raise_on_wrong_dtype(self.fftshift_after, bool, "fftshift_after")
-
-        if (
-            self.ndirs != len(self.nffts)
-            or self.ndirs != len(self.sampling)
-            or self.ndirs != len(self.ifftshift_before)
-            or self.ndirs != len(self.fftshift_after)
-        ):
-            raise ValueError(
-                (
-                    "`dirs`, `nffts`, `sampling`, `ifftshift_before` and "
-                    "`fftshift_after` must the have same number of elements. Received "
-                    f"{self.ndirs}, {len(self.nffts)}, {len(self.sampling)}, "
-                    f"{len(self.ifftshift_before)} and {len(self.fftshift_after)}, "
-                    "respectively."
-                )
-            )
-        self.real = real
-
-        fs = [
-            np.fft.fftshift(np.fft.fftfreq(n, d=s))
-            if fftshift
-            else np.fft.fftfreq(n, d=s)
-            for n, s, fftshift in zip(self.nffts, self.sampling, self.fftshift_after)
-        ]
-        if self.real:
-            fs[-1] = np.fft.rfftfreq(self.nffts[-1], d=self.sampling[-1])
-            if self.fftshift_after[-1]:
-                warnings.warn(
-                    (
-                        "Using real=True and fftshift_after on the last direction. "
-                        "fftshift should only be applied on directions with negative "
-                        "and positive frequencies. When using FFTND with real=True, "
-                        "are all directions except the last. If you wish to proceed "
-                        "applying fftshift on a frequency axis with only positive "
-                        "frequencies, ignore this message."
-                    )
-                )
-                fs[-1] = np.fft.fftshift(fs[-1])
-        self.fs = tuple(fs)
-        self.dims_fft = self.dims.copy()
-        self.dims_fft[self.dirs] = self.nffts
-        if self.real:
-            self.dims_fft[self.dirs[-1]] = self.nffts[-1] // 2 + 1
-        self.shape = (int(np.prod(self.dims_fft)), int(np.prod(self.dims)))
-        self.rdtype = get_real_dtype(dtype) if self.real else np.dtype(dtype)
-        self.cdtype = get_complex_dtype(dtype)
-        self.dtype = self.cdtype
-        self.clinear = False if self.real or np.issubdtype(dtype, np.floating) else True
-        self.explicit = False
-
-    def _matvec(self, x):
-        raise NotImplementedError(
-            "_BaseFFT does not provide _matvec. It must be implemented separately."
-        )
-
-    def _rmatvec(self, x):
-        raise NotImplementedError(
-            "_BaseFFT does not provide _rmatvec. It must be implemented separately."
-        )
+from pylops.signalprocessing._BaseFFTs import _BaseFFTND
 
 
 class FFTND(_BaseFFTND):
